+++ src/pages/_document.tsx
@@ -6,7 +6,6 @@ export default function Document() {
   return (
     <Html lang="en">
       <Head />
-      
       <body>
         <Main />
         <NextScript />
